requirements.txt
XInput-Python
mouse
keyboard
types-keyboard
cx_Freeze
pywin32